prettified files, added prettier script for sol and ts

+++ contracts/external/alpha/ISafeBoxETH.sol
@@ -5,6 +5,8 @@ import "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
 
 interface ISafeBoxETH is IERC20Upgradeable {
   function cToken() external view returns (address);
+
   function deposit() external payable;
-  function withdraw(uint amount) external;
+
+  function withdraw(uint256 amount) external;
 }
