+++ contracts/external/compound/ICErc20.sol
@@ -9,6 +9,11 @@ import "./ICToken.sol";
  * @author Compound
  */
 interface ICErc20 is ICToken {
-    function underlying() external view returns (address);
-    function liquidateBorrow(address borrower, uint repayAmount, ICToken cTokenCollateral) external returns (uint);
+  function underlying() external view returns (address);
+
+  function liquidateBorrow(
+    address borrower,
+    uint256 repayAmount,
+    ICToken cTokenCollateral
+  ) external returns (uint256);
 }
