+++ contracts/external/compound/IPriceOracle.sol
@@ -4,11 +4,11 @@ pragma solidity >=0.8.0;
 import "./ICToken.sol";
 
 interface IPriceOracle {
-    /**
-      * @notice Get the underlying price of a cToken asset
-      * @param cToken The cToken to get the underlying price of
-      * @return The underlying asset price mantissa (scaled by 1e18).
-      *  Zero means the price is unavailable.
-      */
-    function getUnderlyingPrice(ICToken cToken) external view returns (uint);
+  /**
+   * @notice Get the underlying price of a cToken asset
+   * @param cToken The cToken to get the underlying price of
+   * @return The underlying asset price mantissa (scaled by 1e18).
+   *  Zero means the price is unavailable.
+   */
+  function getUnderlyingPrice(ICToken cToken) external view returns (uint256);
 }
