+++ contracts/external/olympus/OlympusStaking.sol
@@ -2,6 +2,7 @@
 pragma solidity >=0.8.0;
 
 abstract contract OlympusStaking {
-    address public OHM;
-    function unstake(uint _amount, bool _trigger) external virtual;
+  address public OHM;
+
+  function unstake(uint256 _amount, bool _trigger) external virtual;
 }
